requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cmake==3.25.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.1
env==0.1.0
filelock==3.6.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.2.0
platformdirs==2.5.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyserial==3.5
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
